@@ -1,14 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.scss';
+import { StrictMode } from 'react';
+import { BrowserRouter } from 'react-router-dom';
 import reportWebVitals from './services/reportWebVitals';
 import App from './App';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root'),
+const container = document.getElementById('app');
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!);
+root.render(
+  <StrictMode>
+    <App router={BrowserRouter} />
+  </StrictMode>,
 );
 
 // If you want to start measuring performance in your app, pass a function
