@@ -1,7 +1,9 @@
-import React from 'react';
-import { useHistory } from 'react-router-dom';
-import { Body, Header, Screen } from '../../components';
-import { withGreeting } from '../../hoc';
+import { FC } from 'react';
+import { useNavigate } from 'react-router-dom';
+import Body from '../../components/Body';
+import Header from '../../components/Header';
+import Screen from '../../components/Screen';
+import withGreeting from '../../hoc/withGreeting';
 import { greet, GreetingProps } from '../../services/greeting';
 
 interface HomeScreenProps extends GreetingProps {
@@ -12,11 +14,11 @@ const defaultProps = {
   componentName: 'HomeScreen',
 };
 
-const HomeScreen : React.FC<HomeScreenProps> = ({ componentName, greeting }) => {
+const HomeScreen : FC<HomeScreenProps> = ({ componentName, greeting }) => {
   greet(greeting, componentName);
 
-  const history = useHistory();
-  history.push('/posts');
+  const navigate = useNavigate();
+  navigate('/posts');
   return (
     <Screen title="Septesting">
       <Header />
