@@ -1,10 +1,10 @@
-import React from 'react';
+import { FC } from 'react';
 import { GreetingProps } from '../../services/greeting';
 import { greeting as greetingConsts } from '../../consts';
 
 function withGreeting<P extends GreetingProps>(Component: React.ComponentType<P>)
-: React.FC<Omit<P, keyof GreetingProps>> {
-  const ComponentWithGreeting = (props: Omit<P, keyof GreetingProps>) => (
+: FC<P> {
+  const ComponentWithGreeting = (props: P) => (
     <Component
     // eslint-disable-next-line react/jsx-props-no-spreading
       {...props as P}
