@@ -1,6 +1,5 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import { BrowserRouter } from 'react-router-dom';
+import { render } from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
 import Component from './PostScreen';
 
 jest.mock('react-router-dom', () => ({
@@ -13,9 +12,9 @@ jest.mock('react-router-dom', () => ({
 test('Checks if component logs rendering', () => {
   console.log = jest.fn();
   render(
-    <BrowserRouter>
+    <MemoryRouter>
       <Component />
-    </BrowserRouter>,
+    </MemoryRouter>,
   );
   expect(console.log).toBeCalled();
 });
