@@ -1,14 +1,13 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import { BrowserRouter } from 'react-router-dom';
+import { render } from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
 import Component from './PostsScreen';
 
 test('Checks if component logs rendering', () => {
   console.log = jest.fn();
   render(
-    <BrowserRouter>
+    <MemoryRouter>
       <Component />
-    </BrowserRouter>,
+    </MemoryRouter>,
   );
   expect(console.log).toBeCalled();
 });
